[REFACTOR] 입력관련 MVC 패턴 변환작업
Exception 처리를 InputView -> Application 으로 수정
